# docs/swagger/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "网络数字孪生平台 REST API 文档。",
        "title": "Network Digital Twin API",
        "contact": {},
        "version": "2.0.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/api/v1/audit": {
            "get": {
                "description": "查询快照操作审计日志",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "audit"
                ],
                "summary": "查询审计日志",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "max items",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "action filter",
                        "name": "action",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "snapshot filter",
                        "name": "snapshot",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
        "/api/v1/snapshot": {
            "get": {
                "description": "返回当前所有快照元数据",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "snapshot"
                ],
                "summary": "列出所有快照",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "post": {
                "description": "创建新的图快照",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "snapshot"
                ],
                "summary": "创建快照",
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/snapshot/diff": {
            "get": {
                "description": "对比两个快照之间的差异",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "snapshot"
                ],
                "summary": "对比快照",
                "parameters": [
                    {
                        "type": "string",
                        "description": "snapshot A name",
                        "name": "a",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "snapshot B name",
                        "name": "b",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/snapshot/restore": {
            "post": {
                "description": "根据名称恢复指定快照",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "snapshot"
                ],
                "summary": "恢复快照",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/snapshot/{name}": {
            "delete": {
                "description": "根据名称删除指定快照",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "snapshot"
                ],
                "summary": "删除快照",
                "parameters": [
                    {
                        "type": "string",
                        "description": "snapshot name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/sync": {
            "post": {
                "description": "从所有 Connector 全量拉取数据并写入 Neo4j",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sync"
                ],
                "summary": "触发全量同步",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.SyncResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/sync/webhook": {
            "post": {
                "description": "接收 Webhook 增量事件，写入事件队列",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sync"
                ],
                "summary": "接收增量事件",
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handlers.SyncResponse": {
            "type": "object",
            "properties": {
                "duration": {
                    "type": "string"
                },
                "nodes_created": {
                    "type": "integer"
                },
                "orphan_edges_skipped": {
                    "type": "integer"
                },
                "relations_created": {
                    "type": "integer"
                }
            }
        },
        "response.ErrorCode": {
            "type": "integer",
            "enum": [
                0,
                400001,
                401001,
                404001,
                429001,
                500001,
                501001,
                503001,
                400011,
                503011,
                400021,
                404021,
                400031,
                500031,
                400041,
                404041,
                501041,
                400051,
                404051,
                501051
            ],
            "x-enum-comments": {
                "CodeBadRequest": "请求参数错误",
                "CodeCircuitBreakOpen": "服务不可用（熔断）",
                "CodeDeviceBadRequest": "设备请求参数错误",
                "CodeDeviceNotFound": "Connector 未找到",
                "CodeDeviceUnsupported": "设备操作不支持",
                "CodeInternalError": "服务内部错误",
                "CodeMonitorBadRequest": "监控请求参数错误",
                "CodeMonitorNotFound": "监控 Connector 未找到",
                "CodeMonitorUnsupported": "监控操作不支持",
                "CodeNotFound": "资源未找到",
                "CodeNotImplemented": "功能未实现",
                "CodeRateLimitExceed": "请求限流",
                "CodeSnapshotBadRequest": "快照请求参数错误",
                "CodeSnapshotNotFound": "快照未找到",
                "CodeSuccess": "成功",
                "CodeSyncQueueFull": "Webhook 队列满",
                "CodeSyncUnsupportedAction": "不支持的同步动作",
                "CodeTopologyBadRequest": "拓扑请求参数错误",
                "CodeTopologyQueryFailed": "拓扑查询失败",
                "CodeUnauthorized": "未认证"
            },
            "x-enum-descriptions": [
                "成功",
                "请求参数错误",
                "未认证",
                "资源未找到",
                "请求限流",
                "服务内部错误",
                "功能未实现",
                "服务不可用（熔断）",
                "不支持的同步动作",
                "Webhook 队列满",
                "快照请求参数错误",
                "快照未找到",
                "拓扑请求参数错误",
                "拓扑查询失败",
                "设备请求参数错误",
                "Connector 未找到",
                "设备操作不支持",
                "监控请求参数错误",
                "监控 Connector 未找到",
                "监控操作不支持"
            ],
            "x-enum-varnames": [
                "CodeSuccess",
                "CodeBadRequest",
                "CodeUnauthorized",
                "CodeNotFound",
                "CodeRateLimitExceed",
                "CodeInternalError",
                "CodeNotImplemented",
                "CodeCircuitBreakOpen",
                "CodeSyncUnsupportedAction",
                "CodeSyncQueueFull",
                "CodeSnapshotBadRequest",
                "CodeSnapshotNotFound",
                "CodeTopologyBadRequest",
                "CodeTopologyQueryFailed",
                "CodeDeviceBadRequest",
                "CodeDeviceNotFound",
                "CodeDeviceUnsupported",
                "CodeMonitorBadRequest",
                "CodeMonitorNotFound",
                "CodeMonitorUnsupported"
            ]
        },
        "response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "业务错误码，0 = 成功",
                    "allOf": [
                        {
                            "$ref": "#/definitions/response.ErrorCode"
                        }
                    ]
                },
                "data": {},
                "message": {
                    "description": "人类可读描述",
                    "type": "string"
                }
            }
        }
    }
}